Initial Commit of His Name is Willard Game
This is the official "His Name is Willard" game, which was built to be
used in a command-line environment. This game is inspired by Will Smith,
who works for Linode, and has the same name (in abbreviation) as Willard
Smith (the actor). This is all for fun. The jokes (mostly) aren't ours.
We will continue to add onto this.

Signed-off-by: Dave Russell <drussell@linode.com>